Okay so I added a butt-load of stuff, read the desc
The player turns around when he walks, the doors open up and the player
walks in, the rift now has an animated texture and it also has a looping
sound. The sounds provided for the buttons were implemented as were the
ones for the cloning machine. I also went around and fixed up every
single level multiple times. The ground cubes now use the layer
"Ground".